btl/uct: add support for using an another memory domain to form connections
In some cases we may have a usable memory domain that does not have a suitable
transport for forming connections. This makes it impossible for the btl, as
written, to work with these networks making the btl unavailable. In many cases
there may be an alternate memory domain and transport (for example, tcp) which
is usable to form connections. This commit changes the setup logic to evaluate
all memory domains and transports (network only) to determine if such a
transport exists. If one does it will be used to form connections only.

DO NOT MERGE -- WIP

Signed-off-by: Nathan Hjelm <hjelmn@google.com>