btl/uct: remove the need for a BTL module for connection TLs
Connection TLs are only used to form connections for connect-to-endpoint TLs.
They do not need to belong to the same memory domain as the one they are used
with so there is no need to rely on a BTL module. This commit moves the
pending_connection_reqs to the tl and changes the code to support a NULL
module for the connection tl.